fix(windows): unbreak src-tauri build on Windows/MSVC
Three independent bugs, only surfaced now that this is the first clean
Windows/MSVC build attempted on this repo:

- browser.rs called `.visible(false)` on `tauri::webview::WebviewBuilder`,
  which has no such method (that's `WebviewWindowBuilder`/`WindowBuilder`
  only) - fails to compile on every platform, not just Windows. Swapped
  for `.hide()` on the built `Webview` handle, matching the "initially
  hidden" contract callers already relied on.

- updater.rs unconditionally imported `std::os::unix::fs::FileExt`, but
  the whole module is a macOS-only DMG updater (mounts DMGs, shells out
  to hdiutil/codesign/spctl/ditto) that was never gated to
  `#[cfg(target_os = "macos")]`. Split into updater_macos.rs (unchanged
  logic) + updater_stub.rs (no-op, same public API) so lib.rs and the
  frontend's invoke() surface stay platform-agnostic.

- tokenizers' default `esaxx_fast` feature builds esaxx-rs's C++ backend
  with a hardcoded static CRT (`.static_crt(true)`), which conflicts with
  the rest of the binary's dynamic CRT (notably `cxx`, via `usearch`) and
  fails MSVC linking with LNK2038. Disabled it in atlas-embed's Cargo.toml
  - it only speeds up training a Unigram tokenizer; loading/running
  pretrained ones is unaffected.
